warning fix

+++ ospray/volume/StructuredVolume.h
@@ -145,7 +145,7 @@ namespace ospray {
 #endif
   template<typename T>
   void StructuredVolume::upsampleRegion(const T *source, T *out, const vec3i &regionSize, const vec3i &scaledRegionSize){
-    for (size_t z = 0; z < scaledRegionSize.z; ++z){
+    for (int z = 0; z < scaledRegionSize.z; ++z){
       parallel_for(scaledRegionSize.x * scaledRegionSize.y, [&](int taskID){
         int x = taskID % scaledRegionSize.x;
         int y = taskID / scaledRegionSize.x;
